Moving DecompressionService
