+++ src/index.ts
@@ -3,5 +3,5 @@ import * as Parser from './service/parser/Parser';
 export { Parser };
 export { default as ChannelApeActionsController } from './channelape/controller/ChannelApeActionsController';
 export { default as SqsMessageService } from './service/sqsMessageService/SqsMessageService';
-export { default as DecompressionService } from './service/decompressionService/DecompressionService';
+export { default as DecompressionService } from './decompression/service/DecompressionService';
 export { default as AdditionalFieldService } from './channelape/service/AdditionalFieldService';
